refactor(stats): extract analytics forecast and support helpers
Move forecast calculations and shared support routines out of the stats analytics facade.

Leave analytics.rs as the public boundary for summary and insight generation.
